Extract run() into run.ts for testability
main.ts executed run() at import time, making the orchestration
logic impossible to test. Pure move; main.ts is now a thin entry.
